Pulled new weapons class and potion class out of the item class file. Removed bow class.
