--- src/main/java/Bow.java
@@ -1,6 +1,0 @@
-
-/*public class Bow extends Weapon {
-public Bow() {
-    super("Bow", 8);
-}
-}*/
