+++ src/main/java/Item.java
@@ -1,0 +1,17 @@
+
+// Base class for all items
+public class Item {
+    protected String name;
+
+    public Item(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void use() {
+        System.out.println("Using " + name);
+    }
+}
