+++ src/main/java/Weapon.java
@@ -1,11 +1,12 @@
 import java.util.Random;
 
-public class Weapon {
+public class Weapon extends Item {
     private int move1MinDamage, move1MaxDamage;
     private int move2MinDamage, move2MaxDamage;
     public String name;
 
     public Weapon(int move1MinDamage, int move1MaxDamage, int move2MinDamage, int move2MaxDamage, String name) {
+        super(name); // Call parent constructor
         this.move1MinDamage = move1MinDamage;
         this.move1MaxDamage = move1MaxDamage;
         this.move2MinDamage = move2MinDamage;
@@ -19,6 +20,11 @@ public class Weapon {
 
     public int getMove2Damage() {
         return new Random().nextInt((move2MaxDamage - move2MinDamage) + 1) + move2MinDamage;
+    }
+
+    @Override
+    public void use() {
+        System.out.println("Equipping " + name);
     }
 
     // Upgrade weapon damage
